Split tests.py into multiple files
For better modularity
