+++ stockhelper/stockapp/tests/utils.py
@@ -1,0 +1,2 @@
+USERNAME = "testuser"
+PASSWORD = "howtostock"
